modified:   components/Card/Card.tsx 	modified:   components/EcosystemApps/EcosystemApps.tsx 	new file:   data/appData.ts 	modified:   pages/user-guides/block-explorers.mdx 	modified:   tailwind.config.js
