+++ components/Card/Card.tsx
@@ -1,5 +1,6 @@
-import Image, { StaticImageData } from "next/image";
+// Card.tsx
 
+import Image, { StaticImageData } from "next/image";
 interface CardProps {
   image: StaticImageData;
   title: string;
@@ -17,10 +18,11 @@ function Card({ image, title, description, href }: CardProps) {
       className="flex flex-col border border-gray-200 dark:border-gray-700 rounded-lg overflow-hidden hover:opacity-80"
       href={href}
       target="_blank"
-      rel="noopener"
+      rel="noopener noreferrer"
     >
       <div>
-        <Image src={image} alt={title} />
+        {/* Assuming 'image' contains a valid path or imported image */}
+        <Image src={image} alt={title} layout="responsive" width={700} height={475} priority />
       </div>
       <div className="flex-1 p-4 bg-gray-100 dark:bg-gray-800">
         <p className="text-lg font-semibold mb-2">{title}</p>
