+++ pages/user-guides/block-explorers.mdx
@@ -2,7 +2,7 @@ import { Tabs } from "nextra/components";
 
 # Block Explorers
 
-## What Is a Block Explorer?**
+## **What Is a Block Explorer?**
 
 Rather than downloading the entire blochain ledger, using a block explorer is the most common method to look up all kinds of useful on-chain information.
 
@@ -17,6 +17,8 @@ Most block explorers offer tools for viewing, and sometimes even interacting wit
 - Network Statistics: Statistics and analytics about the blockchain, such as total number of blocks, transaction throughput, network hash rate, staking and governance data, and much more.
 
 - Smart Contracts: Query smart contract states, deployments and interactions with other wallets or addresses. Some may include the raw code for the contract.
+
+> :warning: **Note**: *In rare cases, some information may be misinterpreted. Off-chain data should be used as a guide only*.
 
 ## Sei Explorers
